Wrote the first component test
